+++ Ending.py
@@ -1,0 +1,15 @@
+# This file was written by Vakaris
+
+def Ending(Status):
+    if (Status == 0):
+        print("Error 404 - NullStatus")
+    elif (Status == 1):
+        print("Congratulations, the winner is Player 1")
+    elif (Status == 2):
+        print("Congratulations, the winner is Player 1")
+    elif (Status == 3):
+        print("This time it's a draw, friendship wins!!!")
+
+# This procedure is used to showcase the result of the game:
+# If someone won - the Winner + no. of moves needed
+# If it was a draw - it was a draw
