+++ Greeting.py
@@ -1,0 +1,16 @@
+# This file was written by Vakaris
+
+import time
+
+def OutputGreeting():
+    Greeting = f"""
+    Welcome to Quarto!
+    In order to win, you have to connect four pieces of idential properties.
+    Best of luck! May the games begin!
+    """
+    print(Greeting)
+    time.sleep(3)
+
+
+# This procedure displays the initial message for the user
+# Also utilizes time to make the experience more enjoyable
