+++ LeftPieces.py
@@ -1,0 +1,10 @@
+# This file was written by Vakaris
+
+import globals
+
+def UpdatePieceList(Piece):
+    globals.LeftPieces.remove(Piece)
+
+
+# This procedure simply updates the list of remaining pieces
+# Takes in a parameter Piece, and then removes it from the global list LeftPieces
