+++ VariableDeclaratiion.py
@@ -1,0 +1,2 @@
+# This file will hold all variables, their type and short description of where they are needed
+#
